backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Jayesh Afre",email = "jayeshafre1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "django (==5.0.6)",
    "djangorestframework (==3.15.2)",
    "djangorestframework-simplejwt (==5.3.1)",
    "django-cors-headers (==4.4.0)",
    "django-environ (==0.11.2)",
    "django-filter (==24.3)",
    "django-redis (==5.4.0)",
    "psycopg2-binary (==2.9.9)",
    "celery (==5.4.0)",
    "redis (==5.0.8)",
    "pillow (==10.4.0)",
    "boto3 (==1.35.0)",
    "django-storages (==1.14.4)",
    "razorpay (==1.4.1)",
    "drf-spectacular (==0.27.2)",
    "gunicorn (==22.0.0)",
    "whitenoise (==6.7.0)",
    "sentry-sdk (==2.13.0)",
    "requests (>=2.33.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (==8.3.2)",
    "pytest-django (==4.8.0)",
    "pytest-cov (==5.0.0)",
    "factory-boy (==3.3.1)",
    "faker (==26.0.0)",
    "black (==24.8.0)",
    "isort (==5.13.2)",
    "flake8 (==7.1.1)",
    "pre-commit (==3.8.0)",
    "django-debug-toolbar (==4.4.6)",
    "ipython (==8.26.0)"
]

[tool.black]
line-length = 88
target-version = ["py311"]
exclude = '''
/(
    \.eggs | \.git | \.hg | \.mypy_cache | \.tox | \.venv
    | _build | buck-out | build | dist | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
skip = ["migrations"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
addopts = "-v --tb=short --cov=apps --cov-report=term-missing"
